Move ConstantSpec, NonNullJSNative and define_constants from utils to interface
