+++ components/script/dom/bindings/codegen/CodegenRust.py
@@ -5353,12 +5353,14 @@ class CGBindingRoot(CGThing):
             'dom::bindings::interface::{NonCallbackInterfaceObjectClass, create_callback_interface_object}',
             'dom::bindings::interface::{create_interface_prototype_object, create_named_constructors}',
             'dom::bindings::interface::{create_noncallback_interface_object, has_instance}',
+            'dom::bindings::interface::{ConstantSpec, NonNullJSNative}',
+            'dom::bindings::interface::ConstantVal::{IntVal, UintVal}',
             'dom::bindings::js::{JS, Root, RootedReference}',
             'dom::bindings::js::{OptionalRootedReference}',
             'dom::bindings::reflector::{Reflectable}',
-            'dom::bindings::utils::{ConstantSpec, DOMClass, DOMJSClass}',
+            'dom::bindings::utils::{DOMClass, DOMJSClass}',
             'dom::bindings::utils::{DOM_PROTO_UNFORGEABLE_HOLDER_SLOT, JSCLASS_DOM_GLOBAL}',
-            'dom::bindings::utils::{NonNullJSNative, ProtoOrIfaceArray, create_dom_global}',
+            'dom::bindings::utils::{ProtoOrIfaceArray, create_dom_global}',
             'dom::bindings::utils::{finalize_global, find_enum_string_index, generic_getter}',
             'dom::bindings::utils::{generic_lenient_getter, generic_lenient_setter}',
             'dom::bindings::utils::{generic_method, generic_setter, get_array_index_from_id}',
@@ -5366,7 +5368,6 @@ class CGBindingRoot(CGThing):
             'dom::bindings::utils::{get_proto_or_iface_array, has_property_on_prototype}',
             'dom::bindings::utils::{is_platform_object, set_dictionary_property}',
             'dom::bindings::utils::{throwing_constructor, trace_global}',
-            'dom::bindings::utils::ConstantVal::{IntVal, UintVal}',
             'dom::bindings::trace::{JSTraceable, RootedTraceable}',
             'dom::bindings::callback::{CallbackContainer,CallbackInterface,CallbackFunction}',
             'dom::bindings::callback::{CallSetup,ExceptionHandling}',
